fix product

+++ app/product/controller.js
@@ -5,33 +5,17 @@ const Product = require("./model.js");
 const Category = require("../category/model.js");
 const Tag = require("../tag/model.js");
 
-/**
- * Handles the storage of a product, including image file processing.
- *
- * This asynchronous function processes an incoming request to store a product.
- * It checks if an image file is included in the request, processes the file by
- * moving it to a target directory, and saves the product information to the database.
- *
- * @param {Object} req - The request object containing product data and optional file.
- * @param {Object} res - The response object used to send back the result.
- * @param {Function} next - The next middleware function in the stack.
- *
- * @returns {Promise<void>} - Returns a JSON response with the saved product or an error message.
- *
- * @throws {Error} - Throws an error if file processing or database operations fail.
- */
 const store = async (req, res, next) => {
   try {
     let payload = req.body;
     console.log("Payload received:", payload);
 
-    //relasi dengan category
     if (payload.category) {
       let category = await Category.findOne({
-        name: { $regex: payload.category, $options: `i` },
+        name: { $regex: payload.category, $options: "i" },
       });
       if (category) {
-        payload = { ...payload, category: category._id };
+        payload.category = category._id;
       } else {
         delete payload.category;
       }
@@ -43,7 +27,7 @@ const store = async (req, res, next) => {
       });
       console.log("Tags found:", tags);
       if (tags.length) {
-        payload = { ...payload, tags: tags.map((tag) => tag._id) };
+        payload.tags = tags.map((tag) => tag._id);
       } else {
         delete payload.tags;
       }
@@ -110,33 +94,29 @@ const update = async (req, res, next) => {
     let payload = req.body;
     let { id } = req.params;
 
-    // Proses category
     if (payload.category) {
       let category = await Category.findOne({
         name: { $regex: payload.category, $options: "i" },
       });
       if (category) {
-        payload = { ...payload, category: category._id };
+        payload.category = category._id;
       } else {
         delete payload.category;
       }
     }
 
-    // Proses tags
     if (payload.tags && payload.tags.length > 0) {
       let tags = await Tag.find({
         _id: { $in: payload.tags },
       });
       if (tags.length) {
-        payload = { ...payload, tags: tags.map((tag) => tag._id) };
+        payload.tags = tags.map((tag) => tag._id);
       } else {
         delete payload.tags;
       }
     }
 
     console.log("Payload received for update:", payload);
-    console.log("Category to update:", payload.category);
-    console.log("Tags to update:", payload.tags);
 
     if (req.file) {
       let tmp_path = req.file.path;
@@ -156,21 +136,28 @@ const update = async (req, res, next) => {
       src.on("end", async () => {
         try {
           let product = await Product.findById(id);
-          let currentImage = `${config.rootPath}/public/images/products/${product.image_url}`;
+          if (!product) {
+            return res
+              .status(404)
+              .json({ error: 1, message: "Product not found" });
+          }
 
+          let currentImage = `${config.rootPath}/public/images/products/${product.image_url}`;
           if (fs.existsSync(currentImage)) {
             fs.unlinkSync(currentImage);
           }
-          product = await Product.findByIdAndUpdate(id, payload, {
-            new: true,
-            runValidators: true,
-          });
-          console.log("Product saved:", product);
+
+          product = await Product.findByIdAndUpdate(
+            id,
+            { ...payload, image_url: filename },
+            { new: true, runValidators: true }
+          );
+          console.log("Product updated:", product);
           return res.json(product);
         } catch (err) {
           fs.unlinkSync(target_path);
           if (err && err.name === "ValidationError") {
-            return res.json({
+            return res.status(400).json({
               error: 1,
               message: err.message,
               fields: err.errors,
@@ -188,12 +175,15 @@ const update = async (req, res, next) => {
         new: true,
         runValidators: true,
       });
-      console.log("Product saved:", product);
+      if (!product) {
+        return res.status(404).json({ error: 1, message: "Product not found" });
+      }
+      console.log("Product updated:", product);
       return res.json(product);
     }
   } catch (err) {
     if (err && err.name === "ValidationError") {
-      return res.json({
+      return res.status(400).json({
         error: 1,
         message: err.message,
         fields: err.errors,
@@ -205,42 +195,45 @@ const update = async (req, res, next) => {
 
 const index = async (req, res, next) => {
   try {
-    let { skip = 0, limit = 10, q = "", category = "", tags = [] } = req.body;
+    let { skip = 0, limit = 10, q = "", category = "", tags = [] } = req.query;
+    skip = parseInt(skip) || 0;
+    limit = parseInt(limit) || 10;
+    let page = parseInt(req.query.page) || 1;
+    console.log("Query page:", page);
+
     let criteria = {};
     if (q.length) {
-      criteria = { ...criteria, name: { $regex: `${q}`, $options: `i` } };
+      criteria.name = { $regex: q, $options: "i" };
     }
 
     if (category.length) {
       let categoryResult = await Category.findOne({
-        name: { $regex: `${category}` },
-        $options: `i`,
+        name: { $regex: category, $options: "i" },
       });
 
       if (categoryResult) {
-        criteria = { ...criteria, category: categoryResult._id };
+        criteria.category = categoryResult._id;
       }
     }
 
     if (tags.length) {
       let tagsResult = await Tag.find({ name: { $in: tags } });
       if (tagsResult.length > 0) {
-        criteria = {
-          ...criteria,
-          tags: { $in: tagsResult.map((tag) => tag._id) },
-        };
+        criteria.tags = { $in: tagsResult.map((tag) => tag._id) };
       }
     }
     console.log(criteria);
-    let count = await Product.find().countDocuments();
-    let product = await Product.find(criteria)
-      .skip(parseInt(skip))
-      .limit(parseInt(limit))
+    let count = await Product.find(criteria).countDocuments();
+    let products = await Product.find(criteria)
+      .skip(skip)
+      .limit(limit)
       .populate("category")
       .populate("tags");
     return res.json({
-      data: product,
-      count,
+      data: products,
+      totalPages: Math.ceil(count / limit),
+      currentPage: page,
+      total: count,
     });
   } catch (err) {
     next(err);
@@ -258,7 +251,6 @@ const destroy = async (req, res, next) => {
       });
     }
 
-    // Hapus gambar jika ada
     if (product.image_url) {
       let currentImage = `${config.rootPath}/public/images/products/${product.image_url}`;
       if (fs.existsSync(currentImage)) {
